Moving new prompting functions into a separate class
