--- src/models/AutoPrompt.ts
@@ -1,4 +1,0 @@
-export interface AutoPromptOptions {
-  force: boolean;
-}
-
